services/requirements.txt
fastapi
crewai
uvicorn
python-multipart
pillow
httpx
python-dotenv
pytest
typing-extensions
